+++ addon/utils.js
@@ -43,6 +43,16 @@ export async function getLatestApiVersionFromOrg(sfHost) {
   }
 }
 
+export async function setOrgInfo(sfHost) {
+  let orgInfo = JSON.parse(sessionStorage.getItem(sfHost + "_orgInfo"));
+  if (orgInfo == null) {
+    const res = await sfConn.rest("/services/data/v" + apiVersion + "/query/?q=SELECT+Id,InstanceName,OrganizationType+FROM+Organization");
+    orgInfo = res.records[0];
+    sessionStorage.setItem(sfHost + "_orgInfo", JSON.stringify(orgInfo));
+  }
+  return orgInfo;
+}
+
 export class PromptTemplate {
   constructor(promptName) {
     this.promptName = promptName;